Create video_player
This is the only place in which code was edited by myself. It was a challenge to set up the Python in order to run the example code and begin editing. But once the editing was accessible and the code runnable I slowly figured out the solutions you see above. Its pretty basic however, I am a beginner to coding and this is my first interactions with object based python.